requirements.txt
Flask
python-dotenv
requests
openai